requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
django-environ==0.12.0
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
python-decouple
requests==2.31.0
huggingface-hub==0.25.1
cairosvg==2.7.1
charset-normalizer==3.4.3
django-crispy-forms==2.1
crispy-bootstrap4==2023.1